Removed broken util get_contract_path
The smart contracts are not part of the raiden code anymore, the
contract manager from the raiden-contracts must be used. This removes
the util get_contract_path since it pointed to a phantom folder, and
changes the utility code to deploy from the contract manager.
